[#67] Refactor: [홈] 목표 데이터 분리, GoalStatus 컴포넌트 매개변수 간소화, GoalProgreeBar 분리
