+++ NeverDie/NeverDie/Modules/Home/View/AddGoalModalView.swift
@@ -7,35 +7,6 @@
 
 import SwiftUI
 
-/// 단계 정보
-struct StageData: Identifiable {
-    let id = UUID()
-    let stageNum: Int
-    let stageContent: String
-}
-
-/// 지표별 데이터
-struct IndicatorData: Identifiable {
-    let id = UUID()
-    let icon: ImageResource
-    let text: String
-    let stages: [StageData]
-}
-
-/// 전체 데이터
-let goalInfoList: [IndicatorData] = [
-    IndicatorData(
-        icon: .stepCountIcon,
-        text: "걸음수",
-        stages: [
-            StageData(stageNum: 1, stageContent: "하루에 4,000걸음 걷기"),
-            StageData(stageNum: 2, stageContent: "하루에 8,000걸음 걷기"),
-            StageData(stageNum: 3, stageContent: "하루에 13,000걸음 걷기")
-        ]
-    )
-    /// 추후 추가 가능
-]
-
 struct AddGoalModalView: View {
     
     // MARK: - Property
